requirements.txt
pytest~=8.3.5
setuptools>=70.0.0
playwright~=1.51.0
requests~=2.32.3
python-dotenv>=1.0.0
